fix: code review fixes and test stability improvements
- Persist constraints to system.constraints, restore on restart
- Fix Raft/TxnManager race: log error but report success after Raft commit
- Add authorization checks before query execution
- Add duplicate key detection for INSERT on user tables
- Fix memtable rotation race with Arc::ptr_eq identity check
- Fix read view creation race by reading max_txn_id under lock
- Add 256MB packet size limit to prevent DoS
- Standardize integer literals to BigInt type
- Remove dead code (empty if-block in typecheck.rs)
- Add input validation for table names, constraints, hex strings
- Improve unsafe block documentation in io_uring code
- Fix test port conflicts by using port 0 for Raft RPC

+++ src/planner/logical/builder.rs
@@ -463,11 +463,11 @@ impl LogicalPlanBuilder {
         match expr {
             ResolvedExpr::Column(col) => col.data_type.clone(),
             ResolvedExpr::Literal(lit) => match lit {
-                Literal::Integer(_) => DataType::Int,
+                Literal::Integer(_) => DataType::BigInt, // Use BigInt for integer literals (safer for large values)
                 Literal::Float(_) => DataType::Double,
                 Literal::String(_) => DataType::Text,
                 Literal::Boolean(_) => DataType::Boolean,
-                Literal::Null => DataType::Int,
+                Literal::Null => DataType::Int, // NULL is polymorphic, default to Int
                 Literal::Blob(_) => DataType::Blob,
             },
             ResolvedExpr::Function { result_type, .. } => result_type.clone(),
